requirements.txt
spacy
pymongo
flask
pymupdf
python-docx